pyproject.toml
[tool.poetry]
name = "dockerview"
version = "0.1.0"
description = "An interactive terminal dashboard for monitoring and managing Docker Compose environments"
authors = ["Chris Dial <dial.chris@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
textual = "^3.3.0"
docker = "^7.0.0"
pyperclip = "^1.9.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
poethepoet = "^0.24.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dockerview = "dockerview.app:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.poe.tasks]
format = [
    {cmd = "black dockerview/"},
    {cmd = "isort dockerview/"}
]
lint = [
    {cmd = "black --check dockerview/"},
    {cmd = "isort --check-only dockerview/"}
]
test = "pytest -v"
check = ["lint", "test"]
all = ["format", "test"]